refactor out a infoCommander
