+++ kifi-backend/modules/cortex/app/com/keepit/controllers/cortex/LDAController.scala
@@ -4,7 +4,7 @@ import com.google.inject.Inject
 import play.api.mvc.Action
 import play.api.libs.json._
 import com.keepit.common.controller.CortexServiceController
-import com.keepit.common.commanders.LDACommander
+import com.keepit.common.commanders.{ LDAInfoCommander, LDACommander }
 import com.keepit.cortex.features.Document
 import com.keepit.cortex.utils.TextUtils
 import com.keepit.cortex.models.lda.{ LDAUserURIInterestScores, LDATopicConfigurations, LDATopicConfiguration, LDATopicInfo }
@@ -14,7 +14,8 @@ import com.keepit.model.{ User, NormalizedURI }
 import com.keepit.common.db.Id
 
 class LDAController @Inject() (
-  lda: LDACommander)
+  lda: LDACommander,
+  infoCommander: LDAInfoCommander)
     extends CortexServiceController {
 
   def numOfTopics() = Action { request =>
@@ -22,8 +23,8 @@ class LDAController @Inject() (
   }
 
   def showTopics(fromId: Int, toId: Int, topN: Int) = Action { request =>
-    val topicWords = lda.topicWords(fromId, toId, topN).map { case (id, words) => (id.toString, words.toMap) }
-    val topicConfigs = lda.topicConfigs(fromId, toId)
+    val topicWords = infoCommander.topicWords(fromId, toId, topN).map { case (id, words) => (id.toString, words.toMap) }
+    val topicConfigs = infoCommander.topicConfigs(fromId, toId)
     val infos = topicWords.map {
       case (tid, words) =>
         val config = topicConfigs(tid)
@@ -53,7 +54,7 @@ class LDAController @Inject() (
   }
 
   def ldaConfigurations = Action { request =>
-    Ok(Json.toJson(lda.ldaConfigurations))
+    Ok(Json.toJson(infoCommander.ldaConfigurations))
   }
 
   def getLDAFeatures() = Action.async(parse.tolerantJson) { request =>
